+++ CHANGELOG.rst
@@ -1,6 +1,15 @@
 Changelog
 =========
 
+Release *v2.0.dev0* - ``2019-10-07``
+------------------------------------
+* Support for redis-py >= 3 only
+* Support for redis-server >= 3 only
+* Breaking change: `zadd` value/scores cannot be passed as positional arguments anymore
+* Breaking change: `zadd` flags other than `ch` are explicitely not supported
+* Breaking change: `zincrby` arguments are swaped (`amount, value` instead of `value, amount`)
+* Breaking change: Redis server with `LUA` scripting support is mandatory
+
 Release *v1.3.1* - ``2019-10-11``
 ---------------------------------
 * Resolve race condition in `get_or_connect`